Rename icons in BottomNavigator component to new icon components
